+++ manifests/memcache-operator.yaml
@@ -19,7 +19,7 @@ spec:
       serviceAccountName: memcached-operator
       containers:
         - name: memcached-operator
-          image: mypathinmyheart/memcached-operator:0.1alpha
+          image: mypathinmyheart/memcached-operator:1alpha
           imagePullPolicy: Always
           resources:
             requests:
