+++ manifests/operator-service-account-rbac.yaml
@@ -39,6 +39,7 @@ rules:
       - apps
     resources:
       - statefulsets
+      - deployments
     verbs:
       - get
       - list
